+++ Core/modules.py
@@ -1,4 +1,5 @@
 from OpenGL.GLUT import glutInit, glutBitmapCharacter, GLUT_BITMAP_HELVETICA_12 # type: ignore
+from cryptography.fernet import Fernet
 from OpenGL import GL
 from enum import Enum
 import keyboard
@@ -10,3 +11,5 @@ import os
 import psutil
 import threading
 import random
+import sys
+import json